+++ src/modules/Servicehosting/Api/Admin.php
@@ -227,6 +227,14 @@ class Admin extends \Api_Abstract
 
         $model = $this->di['db']->getExistingModelById('ServiceHostingServer', $data['id'], 'Server not found');
 
+        // check if server is not used by any service_hostings
+        $hosting_services = $this->di['db']->find('ServiceHosting', 'service_hosting_server_id = :cart_id', [':cart_id' => $data['id']]);
+        $count = is_array($hosting_services) ? count($hosting_services) : 0; // Handle the case where $hosting_services might be null
+
+        if ($count > 0) {
+            throw new \Box_Exception('Hosting server is used by :count: service hostings', [':count:' => $count], 704);
+        }
+
         return (bool) $this->getService()->deleteServer($model);
     }
 
@@ -250,6 +258,7 @@ class Admin extends \Api_Abstract
      *
      * @throws \Box_Exception
      */
+
     public function server_update($data)
     {
         $required = [
@@ -330,8 +339,18 @@ class Admin extends \Api_Abstract
 
         $model = $this->di['db']->getExistingModelById('ServiceHostingHp', $data['id'], 'Hosting plan not found');
 
+        // check if hosting plan is not used by any service_hostings
+        $hosting_services = $this->di['db']->find('ServiceHosting', 'service_hosting_hp_id = :cart_id',[':cart_id' => $data['id']]);
+
+        // Ensure $hosting_services is an array before counting its elements
+        $count = is_array($hosting_services) ? count($hosting_services) : 0; // Handle the case where $hosting_services might be null
+        if ($count > 0)
+        {
+            throw new \Box_Exception('Hosting plan is used by :count: service hostings', [':count:'=> $count], 704);
+        }
         return (bool) $this->getService()->deleteHp($model);
     }
+
 
     /**
      * Get hosting plan details.
